moderator assignment

+++ app/controllers/posts_controller.rb
@@ -42,6 +42,11 @@ class PostsController < ApplicationController
    end
 
   def destroy
+    if current_user.moderator?
+      flash[:alert] = 'You must be an admin to do that.'
+      redirect_to [post.topic, post]
+    end
+
     @post = Post.find(params[:id])
 
     if @post.destroy
@@ -61,7 +66,7 @@ class PostsController < ApplicationController
 
   def authorize_user
     post = Post.find(params[:id])
-    unless current_user == post.user || current_user.admin?
+    unless current_user == post.user || current_user.admin? || current_user.moderator?
       flash[:alert] = 'You must be an admin to do that.'
       redirect_to [post.topic, post]
     end
