+++ app/controllers/topics_controller.rb
@@ -15,6 +15,11 @@ class TopicsController < ApplicationController
   end
 
   def create
+    if current_user.moderator?
+      flash[:alert] = 'You must be an admin to do that.'
+      redirect_to topics_path
+    end
+
     @topic = Topic.new(topic_params)
 
     if @topic.save
@@ -42,6 +47,11 @@ class TopicsController < ApplicationController
    end
 
   def destroy
+    if current_user.moderator?
+      flash[:alert] = 'You must be an admin to do that.'
+      redirect_to topics_path
+    end
+
     @topic = Topic.find(params[:id])
 
     if @topic.destroy
@@ -60,7 +70,7 @@ class TopicsController < ApplicationController
   end
 
   def authorize_user
-    unless current_user.admin?
+    unless current_user.admin? || current_user.moderator?
       flash[:alert] = 'You must be an admin to do that.'
       redirect_to topics_path
     end
